renames abstracts to abstract_communicator.py renames "acl" to "icl"
