+++ horiba_sdk/communication/__init__.py
@@ -28,7 +28,7 @@ await ws.disconnect()
 """
 
 # Necessary to make Python treat the directory as a package
-from .abstracts import AbstractCommunicator
+from .abstract_communicator import AbstractCommunicator
 from .websocket_communicator import WebsocketCommunicator
 
 __all__ = ['AbstractCommunicator', 'WebsocketCommunicator']
