+++ horiba_sdk/communication/abstract_communicator.py
